OvmfPkg/AcpiPlatformDxe: Fix VS2010 build issue
Reported-by: Liming Gao <liming.gao@intel.com>
Contributed-under: TianoCore Contribution Agreement 1.0

git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/trunk/edk2@13425 6f19259b-4bc3-4df7-8a09-765794883524

+++ OvmfPkg/AcpiPlatformDxe/Qemu.c
@@ -82,13 +82,13 @@ QemuInstallAcpiMadtTable (
   for (Loop = 1; Loop < Count; Loop++) {
     LocalApic->Type = EFI_ACPI_1_0_PROCESSOR_LOCAL_APIC;
     LocalApic->Length = sizeof (*LocalApic);
-    LocalApic->AcpiProcessorId = Loop;
-    LocalApic->ApicId = Loop;
+    LocalApic->AcpiProcessorId = (UINT8) Loop;
+    LocalApic->ApicId = (UINT8) Loop;
     LocalApic->Flags = 1;
     LocalApic++;
   }
 
-  Hdr->Length = NewBufferSize;
+  Hdr->Length = (UINT32) NewBufferSize;
 
   Status = InstallAcpiTable (AcpiProtocol, Hdr, NewBufferSize, TableKey);
